+++ db/facelog-db-base/src/main/java/net/gdface/facelog/db/FaceBean.java
@@ -95,11 +95,11 @@ public final class FaceBean
     /** flag whether {@code this} can be modified */
     private Boolean immutable;
     /** columns modified flag */
-    @ApiModelProperty(value="columns modified flag",dataType="long",required=true)
-    private long modified;
+    @ApiModelProperty(value="columns modified flag",dataType="int",required=true)
+    private int modified;
     /** columns initialized flag */
-    @ApiModelProperty(value="columns initialized flag",dataType="long",required=true)
-    private long initialized;
+    @ApiModelProperty(value="columns initialized flag",dataType="int",required=true)
+    private int initialized;
     /** new record flag  */
     @ApiModelProperty(value="new record flag",dataType="boolean",required=true)
     private boolean isNew;        
@@ -167,7 +167,7 @@ public final class FaceBean
      * @return the modified status of columns
      */
     @ThriftField(value=2,requiredness=Requiredness.REQUIRED)
-    public long getModified(){
+    public int getModified(){
         return modified;
     }
 
@@ -175,14 +175,14 @@ public final class FaceBean
      * @param modified the modified status bit to be assigned to {@link #modified}
      */
     @ThriftField()
-    public void setModified(long modified){
+    public void setModified(int modified){
         this.modified = modified;
     }
     /**
      * @return the initialized status of columns
      */
     @ThriftField(value=3,requiredness=Requiredness.REQUIRED)
-    public long getInitialized(){
+    public int getInitialized(){
         return initialized;
     }
 
@@ -190,7 +190,7 @@ public final class FaceBean
      * @param initialized the initialized status bit to be assigned to {@link #initialized}
      */
     @ThriftField()
-    public void setInitialized(long initialized){
+    public void setInitialized(int initialized){
         this.initialized = initialized;
     }
     protected static final <T extends Comparable<T>>boolean equals(T a, T b) {
@@ -1793,7 +1793,7 @@ public final class FaceBean
     public void resetIsModified()
     {
         checkMutable();
-        modified = 0L;
+        modified = 0;
     }
 
     @Override
@@ -1830,7 +1830,7 @@ public final class FaceBean
      */
     private void resetInitialized()
     {
-        initialized = 0L;
+        initialized = 0;
     }
     /** reset all fields to initial value, equal to a new bean */
     public void reset(){
@@ -1855,8 +1855,8 @@ public final class FaceBean
         this.extInfo = null;
         this.featureMd5 = null;
         this.isNew = true;
-        this.modified = 0L;
-        this.initialized = 0L;
+        this.modified = 0;
+        this.initialized = 0;
     }
     @Override
     public boolean equals(Object object)
@@ -2196,7 +2196,7 @@ public final class FaceBean
      * @see #asNULL()
      */
     public boolean checkNULL(){
-        return 0L == getInitialized();
+        return 0 == getInitialized();
     }
     /** 
      * @param source source list
