+++ db/facelog-db-orm/src/main/java/net/gdface/facelog/dborm/device/FlDeviceBean.java
@@ -78,9 +78,9 @@ public  class FlDeviceBean
     /** flag whether {@code this} can be modified */
     private Boolean immutable;
     /** columns modified flag */
-    private long modified;
+    private int modified;
     /** columns initialized flag */
-    private long initialized;
+    private int initialized;
     /** new record flag  */
     private boolean isNew;        
     /** 
@@ -144,27 +144,27 @@ public  class FlDeviceBean
     /**
      * @return the modified status of columns
      */
-    public long getModified(){
+    public int getModified(){
         return modified;
     }
 
     /**
      * @param modified the modified status bit to be assigned to {@link #modified}
      */
-    public void setModified(long modified){
+    public void setModified(int modified){
         this.modified = modified;
     }
     /**
      * @return the initialized status of columns
      */
-    public long getInitialized(){
+    public int getInitialized(){
         return initialized;
     }
 
     /**
      * @param initialized the initialized status bit to be assigned to {@link #initialized}
      */
-    public void setInitialized(long initialized){
+    public void setInitialized(int initialized){
         this.initialized = initialized;
     }
     protected static final <T extends Comparable<T>>boolean equals(T a, T b) {
@@ -1234,7 +1234,7 @@ public  class FlDeviceBean
     public void resetIsModified()
     {
         checkMutable();
-        modified = 0L;
+        modified = 0;
     }
 
     @Override
@@ -1268,7 +1268,7 @@ public  class FlDeviceBean
      */
     private void resetInitialized()
     {
-        initialized = 0L;
+        initialized = 0;
     }
     /** reset all fields to initial value, equal to a new bean */
     public void reset(){
@@ -1293,7 +1293,7 @@ public  class FlDeviceBean
         /* DEFAULT:'CURRENT_TIMESTAMP'*/
         this.updateTime = null;
         this.isNew = true;
-        this.modified = 0L;
+        this.modified = 0;
         this.initialized = (FL_DEVICE_ID_GROUP_ID_MASK);
     }
     @Override
@@ -1598,7 +1598,7 @@ public  class FlDeviceBean
      * @see #asNULL()
      */
     public boolean checkNULL(){
-        return 0L == getInitialized();
+        return 0 == getInitialized();
     }
     /** 
      * @param source source list
