+++ db/facelog-db-orm/src/main/java/net/gdface/facelog/dborm/image/FlImageBean.java
@@ -57,9 +57,9 @@ public  class FlImageBean
     /** flag whether {@code this} can be modified */
     private Boolean immutable;
     /** columns modified flag */
-    private long modified;
+    private int modified;
     /** columns initialized flag */
-    private long initialized;
+    private int initialized;
     /** new record flag  */
     private boolean isNew;        
     /** 
@@ -123,27 +123,27 @@ public  class FlImageBean
     /**
      * @return the modified status of columns
      */
-    public long getModified(){
+    public int getModified(){
         return modified;
     }
 
     /**
      * @param modified the modified status bit to be assigned to {@link #modified}
      */
-    public void setModified(long modified){
+    public void setModified(int modified){
         this.modified = modified;
     }
     /**
      * @return the initialized status of columns
      */
-    public long getInitialized(){
+    public int getInitialized(){
         return initialized;
     }
 
     /**
      * @param initialized the initialized status bit to be assigned to {@link #initialized}
      */
-    public void setInitialized(long initialized){
+    public void setInitialized(int initialized){
         this.initialized = initialized;
     }
     protected static final <T extends Comparable<T>>boolean equals(T a, T b) {
@@ -711,7 +711,7 @@ public  class FlImageBean
     public void resetIsModified()
     {
         checkMutable();
-        modified = 0L;
+        modified = 0;
     }
 
     @Override
@@ -737,7 +737,7 @@ public  class FlImageBean
      */
     private void resetInitialized()
     {
-        initialized = 0L;
+        initialized = 0;
     }
     /** reset all fields to initial value, equal to a new bean */
     public void reset(){
@@ -753,7 +753,7 @@ public  class FlImageBean
         this.thumbMd5 = null;
         this.deviceId = null;
         this.isNew = true;
-        this.modified = 0L;
+        this.modified = 0;
         this.initialized = (FL_IMAGE_ID_DEPTH_MASK | FL_IMAGE_ID_FACE_NUM_MASK);
     }
     @Override
@@ -962,7 +962,7 @@ public  class FlImageBean
      * @see #asNULL()
      */
     public boolean checkNULL(){
-        return 0L == getInitialized();
+        return 0 == getInitialized();
     }
     /** 
      * @param source source list
