+++ db/facelog-db-orm/src/main/java/net/gdface/facelog/dborm/person/FlPersonManager.java
@@ -280,7 +280,7 @@ public class FlPersonManager extends TableManager.BaseAdapter<FlPersonBean>
         if(null == bean  || null == bean.getId()){
             return false;
         }
-        long modified = bean.getModified();
+        int modified = bean.getModified();
         try{
             bean.resetModifiedExceptPrimaryKeys();
             return 1 == countUsingTemplate(bean);
